app/requirements.txt
pillow==6.2.0
tensorflow==2.0.0
streamlit==0.79.0
pandas==0.25.2
numpy==1.19.5
